Added cross-categories transactions, form, buttons, displaying and tests

+++ spec/factories/categories.rb
@@ -1,7 +1,7 @@
 FactoryBot.define do
   factory :category do
     name { Faker::Lorem.word }
-    amount { Faker::Number.digit.to_i }
+    amount { Faker::Number.digit.to_f }
     user
   end
 end
